+++ collections/collections.d.ts
@@ -9,3 +9,4 @@ export function createCollection({ name, schema, collection, helpers, apply, com
     options?: any;
 }): Mongo.Collection;
 export { persistable } from "./composers/persistable";
+export { softRemoval } from "./composers/softRemoval";