+++ collections/composers/softRemoval.d.ts
@@ -1,0 +1,1 @@
+export function softRemoval(collection: Mongo.Collection): any;