+++ collections/CHANGELOG.md
@@ -3,6 +3,7 @@
 ### 3.0.2 (2024-10-10)
 
 - Adds `persistable` composer
+- Adds `softRemoval` composer
 
 ### 3.0.1 (2024-10-10)